chore: Reorganize app config
