+++ src/service/src/models/mod.rs
@@ -3,6 +3,7 @@
 // Copyright: 2024–2025, Rémi Bardon <remi@remibardon.name>
 // License: Mozilla Public License v2.0 (MPL v2.0)
 
+mod atoms;
 mod avatar;
 mod color;
 mod data;
@@ -16,17 +17,18 @@ mod serializable_secret_string;
 mod url;
 pub mod xmpp;
 
-pub use avatar::*;
-pub use color::*;
-pub use data::*;
-pub use durations::*;
-pub use email_address::*;
-pub use lua::*;
-pub use paginated::*;
-pub use pagination::*;
-pub use serializable_secret_string::*;
-pub use url::*;
-pub use xmpp::{
+pub use self::atoms::*;
+pub use self::avatar::*;
+pub use self::color::*;
+pub use self::data::*;
+pub use self::durations::*;
+pub use self::email_address::*;
+pub use self::lua::*;
+pub use self::paginated::*;
+pub use self::pagination::*;
+pub use self::serializable_secret_string::*;
+pub use self::url::*;
+pub use self::xmpp::{
     jid, BareJid, FullJid, JidDomain, JidNode, XmppConnectionType, XmppDirectionalConnectionType,
     JID,
 };
